refactor: make naming of components consistent
